remove samples folder
