--- samples/library/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-bin
-lib
-dist
-.DS_Store
