--- samples/library/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "sdk-library",
-  "version": "1.0.0",
-  "description": "Dcl SDK Library",
-  "main": "./dist/index.js",
-  "scripts": {
-    "build": "rollup -c --environment BUILD:production",
-    "link": "npm link && rollup --watch -c --environment BUILD:production"
-  },
-  "typings": "./dist/index.d.ts",
-  "decentralandLibrary": {
-    "main": "./dist/index.js"
-  },
-  "keywords": [
-    "dcl",
-    "decentraland",
-    "sdk",
-    "library"
-  ],
-  "author": "-",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/$author/$name"
-  },
-  "files": [
-    "dist"
-  ],
-  "devDependencies": {
-    "@rollup/plugin-typescript": "^8.3.1",
-    "@rollup/plugin-commonjs": "^21.0.2",
-    "@rollup/plugin-node-resolve": "^13.1.3",
-    "decentraland-ecs": "latest",
-    "rollup": "^2.70.1",
-    "rollup-plugin-terser": "^7.0.2"
-  },
-  "dependencies": {
-  },
-  "engines": {
-    "node": ">=16.0.0",
-    "npm": ">=8.0.0",
-    "yarn": "please use npm"
-  }
-}
