--- samples/library/src/exampleComponent.ts
@@ -1,5 +1,0 @@
-@Component('exampleComponent')
-export class ExampleComponent {
-  constructor(public testing: boolean) {
-  }
-}
