--- samples/library/src/index.ts
@@ -1,1 +1,0 @@
-export { ExampleComponent } from './exampleComponent'
