--- samples/library/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": [
-    "dist"
-  ],
-  "compilerOptions": {
-    "module": "ESNext",
-    "removeComments": true,
-    "declaration": true,
-    "declarationDir": ".",
-    "outDir": "dist"
-  },
-  "extends": "./node_modules/decentraland-ecs/types/tsconfig.json"
-}
