--- samples/scene/Dockerfile
@@ -1,21 +1,0 @@
-FROM node:lts-alpine
-
-# Create app directory
-WORKDIR /usr/src/app
-
-RUN npm install --global decentraland@next
-
-# Install app dependencies
-# A wildcard is used to ensure both package.json AND package-lock.json are copied
-# where available (npm@5+)
-COPY package*.json ./
-
-RUN npm install
-# If you are building your code for production
-# RUN npm install --only=production
-
-# Bundle app source
-COPY . .
-
-EXPOSE 8000
-CMD [ "npm", "start", "--", "--ci" ]
