--- samples/scene/eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "extends": "@dcl/eslint-config/sdk",
-  "parserOptions": {
-    "project": ["tsconfig.json"]
-  }
-}
