--- samples/scene/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "dcl-project",
-  "version": "1.0.0",
-  "description": "My new Decentraland project",
-  "scripts": {
-    "start": "dcl start",
-    "build": "build-ecs",
-    "watch": "build-ecs --watch",
-    "deploy:now": "dcl export && now export",
-    "ecs:install": "npm install --save-dev decentraland-ecs@latest",
-    "ecs:install-next": "npm install --save-dev decentraland-ecs@next",
-    "lint": "eslint . --ext .ts",
-    "lint:fix": "eslint . --ext .ts --fix"
-  },
-  "devDependencies": {
-    "@dcl/eslint-config": "^1.0.1",
-    "decentraland-ecs": "latest"
-  },
-  "engines": {
-    "node": ">=14.0.0",
-    "npm": ">=6.0.0",
-    "yarn": "please use npm"
-  }
-}
