--- samples/scene/scene.json
@@ -1,37 +1,0 @@
-{
-  "display": {
-    "title": "DCL Scene",
-    "description": "My new Decentraland project",
-    "navmapThumbnail": "images/scene-thumbnail.png",
-    "favicon": "favicon_asset"
-  },
-  "owner": "",
-  "contact": {
-    "name": "author-name",
-    "email": ""
-  },
-  "main": "bin/game.js",
-  "tags": [],
-  "scene": {
-    "parcels": ["0,0"],
-    "base": "0,0"
-  },
-  "spawnPoints": [
-    {
-      "name": "spawn1",
-      "default": true,
-      "position": {
-        "x": 0,
-        "y": 0,
-        "z": 0
-      },
-      "cameraTarget": {
-        "x": 8,
-        "y": 1,
-        "z": 8
-      }
-    }
-  ],
-  "requiredPermissions": [],
-  "featureToggles": {}
-}
