--- samples/scene/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "outFile": "./bin/game.js",
-    "allowJs": true,
-    "strict": true
-  },
-  "include": ["src/**/*.ts"],
-  "extends": "./node_modules/decentraland-ecs/types/tsconfig.json"
-}
