+++ typings/dcl.d.ts
@@ -1,7 +1,3 @@
-declare module '*.json' {
-  export const abi: any
-}
-
 declare module 'wildcards' {
   const wildcards: (...args: unknown) => void
   export = wildcards
